Refactor and to add markers on map
