+++ Packages/CustomMap/Sources/CustomMap/GoogleMaps/Entity/CustomMapCoordinates.swift
@@ -11,12 +11,3 @@ struct CustomMapCoordinates {
     let latitude: Double
     let longitude: Double
 }
-
-extension CustomMapCoordinates {
-    var toCLLocationCoordinate2D: CLLocationCoordinate2D {
-        CLLocationCoordinate2D(
-            latitude: latitude,
-            longitude: longitude
-        )
-    }
-}
