+++ Packages/CustomMap/Sources/CustomMap/GoogleMaps/Entity/MarkerInfo.swift
@@ -1,10 +1,11 @@
 //
-//  File.swift
+//  MarkerInfo.swift
 //  
 //
 //  Created by Andres Felipe Alzate Restrepo on 30/10/22.
 //
 
 struct MarkerInfo {
-    let indicator: String
+    let indicator: String?
+    let position: CustomMapCoordinates
 }
