+++ Packages/CustomMap/Sources/CustomMap/GoogleMaps/Extension/CustomMapCoordinates+Coordinates.swift
@@ -1,0 +1,15 @@
+//
+//  CustomMapCoordinates+Coordinates.swift
+//  
+//
+//  Created by Andres Felipe Alzate Restrepo on 30/10/22.
+//
+
+extension CustomMapCoordinates {
+    init(_ coordinates: Coordinates) {
+        self.init(
+            latitude: coordinates.latitude,
+            longitude: coordinates.longitude
+        )
+    }
+}
