+++ Packages/CustomMap/Sources/CustomMap/GoogleMaps/Extension/MarkerInfo+GMSMarker.swift
@@ -1,0 +1,14 @@
+//
+//  MarkerInfo+GMSMarker.swift
+//  
+//
+//  Created by Andres Felipe Alzate Restrepo on 30/10/22.
+//
+
+import GoogleMaps
+
+extension MarkerInfo {
+    var asGMSMarker: GMSMarker {
+        GMSMarker(position: position.toCLLocationCoordinate2D)
+    }
+}
